+++ lib/core/router/app_router.dart
@@ -13,12 +13,12 @@ import 'package:cherrypic/presentation/screens/my_page/notice/notice_screen.dart
 import 'package:cherrypic/presentation/screens/my_page/settings/setting_screen.dart';
 import 'package:cherrypic/presentation/screens/my_page/album_payment_info/album_payment_info_screen.dart';
 import 'package:cherrypic/presentation/screens/store/photo_printing/photo_printing_screen.dart';
-import 'package:cherrypic/presentation/screens/store/photo_printing/service_step/change_address_screen.dart';
-import 'package:cherrypic/presentation/screens/store/photo_printing/service_step/select_address_screen.dart';
-import 'package:cherrypic/presentation/screens/store/photo_printing/service_step/select_album_screen.dart';
-import 'package:cherrypic/presentation/screens/store/photo_printing/service_step/select_image_screen.dart';
-import 'package:cherrypic/presentation/screens/store/photo_printing/service_step/select_option_screen.dart';
-import 'package:cherrypic/presentation/screens/store/photo_printing/service_step/select_payment_screen.dart';
+import 'package:cherrypic/presentation/screens/store/photo_printing/service_step/select_address/change_address/change_address_screen.dart';
+import 'package:cherrypic/presentation/screens/store/photo_printing/service_step/select_address/select_address_screen.dart';
+import 'package:cherrypic/presentation/screens/store/photo_printing/service_step/select_album/select_album_screen.dart';
+import 'package:cherrypic/presentation/screens/store/photo_printing/service_step/select_image/select_image_screen.dart';
+import 'package:cherrypic/presentation/screens/store/photo_printing/service_step/select_option/select_option_screen.dart';
+import 'package:cherrypic/presentation/screens/store/photo_printing/service_step/select_payment/select_payment_screen.dart';
 import 'package:cherrypic/presentation/screens/store/subscription/payment_complete_screen.dart';
 import 'package:cherrypic/presentation/screens/store/subscription/payment_method_screen.dart';
 import 'package:cherrypic/presentation/screens/store/store_main_screen.dart';
