+++ lib/presentation/screens/store/photo_printing/service_step/select_address/change_address/change_address_view_model.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../my_page/address_management/address_management_screen.dart';
+import '../../../../../my_page/address_management/address_management_screen.dart';
 
 class ChangeAddressViewModel extends ChangeNotifier {
   final List<AddressItem> _addressItems = [];
