+++ lib/presentation/screens/store/photo_printing/service_step/select_payment/payment/print_payment_info.dart
@@ -1,10 +1,10 @@
-import 'package:cherrypic/presentation/screens/store/photo_printing/payment/print_payment_info_model.dart';
-import 'package:cherrypic/presentation/screens/store/photo_printing/payment/print_payment_info_view_model.dart';
+import 'package:cherrypic/presentation/screens/store/photo_printing/service_step/select_payment/payment/print_payment_info_model.dart';
+import 'package:cherrypic/presentation/screens/store/photo_printing/service_step/select_payment/payment/print_payment_info_view_model.dart';
 import 'package:flutter/material.dart';
 import 'package:cherrypic/core/constants/color.dart';
 import 'package:cherrypic/core/constants/font.dart';
 
-import '../../../../widgets/custom_box_card.dart';
+import '../../../../../../widgets/custom_box_card.dart';
 
 /// 인화 결제 정보
 class PrintPaymentInfo extends StatelessWidget {
